--- src/main/java/org/example/flowday/domain/post/controller/PostController.java
@@ -1,57 +1,0 @@
-package org.example.flowday.domain.post.controller;
-
-import jakarta.validation.Valid;
-import lombok.RequiredArgsConstructor;
-import org.example.flowday.domain.post.dto.PostRequestDTO;
-import org.example.flowday.domain.post.dto.PostResponseDTO;
-import org.example.flowday.domain.post.service.PostService;
-import org.springframework.http.HttpStatus;
-import org.springframework.http.ResponseEntity;
-import org.springframework.web.bind.annotation.*;
-
-import java.util.List;
-import java.util.Optional;
-
-@RestController
-@RequestMapping("/api/posts")
-@RequiredArgsConstructor
-public class PostController {
-
-    private final PostService postService;
-
-    // 게시글 생성
-    @PostMapping
-    public ResponseEntity<PostResponseDTO> createPost(@Valid @RequestBody PostRequestDTO postRequestDto) {
-        PostResponseDTO createdPost = postService.createPost(postRequestDto);
-        return new ResponseEntity<>(createdPost, HttpStatus.CREATED);
-    }
-
-    // 게시글 단건 조회
-    @GetMapping("/{id}")
-    public ResponseEntity<PostResponseDTO> getPostById(@PathVariable Long id) {
-        Optional<PostResponseDTO> post = postService.getPostById(id);
-        return post.map(value -> new ResponseEntity<>(value, HttpStatus.OK))
-                .orElseGet(() -> new ResponseEntity<>(HttpStatus.NOT_FOUND));
-    }
-
-    // 모든 게시글 조회
-    @GetMapping
-    public ResponseEntity<List<PostResponseDTO>> getAllPosts() {
-        List<PostResponseDTO> posts = postService.getAllPosts();
-        return new ResponseEntity<>(posts, HttpStatus.OK);
-    }
-
-    // 게시글 수정
-    @PutMapping("/{id}")
-    public ResponseEntity<PostResponseDTO> updatePost(@PathVariable Long id, @Valid @RequestBody PostRequestDTO updatedPostDto) {
-        PostResponseDTO post = postService.updatePost(id, updatedPostDto);
-        return new ResponseEntity<>(post, HttpStatus.OK);
-    }
-
-    // 게시글 삭제
-    @DeleteMapping("/{id}")
-    public ResponseEntity<Void> deletePost(@PathVariable Long id) {
-        postService.deletePost(id);
-        return new ResponseEntity<>(HttpStatus.NO_CONTENT);
-    }
-}
