--- src/main/java/org/example/flowday/domain/post/dto/PostResponseDTO.java
@@ -1,21 +1,0 @@
-package org.example.flowday.domain.post.dto;
-
-import lombok.*;
-
-import java.time.LocalDateTime;
-
-@Getter
-@Setter
-@NoArgsConstructor
-@AllArgsConstructor
-@Builder
-public class PostResponseDTO {
-    private Long id;
-    private Long memberId;
-    private String city;
-    private String title;
-    private String contents;
-    private LocalDateTime createdAt;
-    private LocalDateTime updatedAt;
-    private Long courseId;
-}
