--- src/main/java/org/example/flowday/domain/post/entity/Status.java
@@ -1,5 +1,0 @@
-package org.example.flowday.domain.post.entity;
-
-public enum Status {
-    PRIVATE , PUBLIC , COUPLE
-}
