+++ src/main/java/org/example/flowday/domain/post/likes/dto/LikeRequestDTO.java
@@ -1,0 +1,17 @@
+package org.example.flowday.domain.post.likes.dto;
+
+import lombok.*;
+import jakarta.validation.constraints.NotNull;
+
+@Getter
+@Setter
+@NoArgsConstructor
+@AllArgsConstructor
+@Builder
+public class LikeRequestDTO {
+    @NotNull(message = "사용자 ID는 필수입니다.")
+    private Long userId;
+
+    @NotNull(message = "게시물 ID는 필수입니다.")
+    private Long postId;
+}
