+++ src/main/java/org/example/flowday/domain/post/post/entity/Status.java
@@ -1,0 +1,5 @@
+package org.example.flowday.domain.post.post.entity;
+
+public enum Status {
+    PRIVATE , PUBLIC , COUPLE
+}
