--- src/main/java/org/example/flowday/domain/post/repository/PostRepository.java
@@ -1,9 +1,0 @@
-package org.example.flowday.domain.post.repository;
-
-import org.example.flowday.domain.post.entity.Post;
-import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.stereotype.Repository;
-
-@Repository
-public interface PostRepository extends JpaRepository<Post, Long> {
-}
