+++ src/main/java/org/example/flowday/domain/post/tag/repository/TagRepository.java
@@ -1,0 +1,10 @@
+package org.example.flowday.domain.post.tag.repository;
+
+import org.example.flowday.domain.post.tag.entity.Tag;
+import org.springframework.data.jpa.repository.JpaRepository;
+import org.springframework.stereotype.Repository;
+
+@Repository
+public interface TagRepository extends JpaRepository<Tag, Long> {
+    boolean existsByName(String name);
+}
