--- src/main/java/org/example/flowday/domain/tag/dto/TagRequestDTO.java
@@ -1,14 +1,0 @@
-package org.example.flowday.domain.tag.dto;
-
-import jakarta.validation.constraints.NotBlank;
-import lombok.*;
-
-@Getter
-@Setter
-@NoArgsConstructor
-@AllArgsConstructor
-@Builder
-public class TagRequestDTO {
-    @NotBlank(message = "태그 이름은 필수입니다.")
-    private String name;
-}
