--- src/main/java/org/example/flowday/domain/tag/repository/TagRepository.java
@@ -1,12 +1,0 @@
-package org.example.flowday.domain.tag.repository;
-
-import org.example.flowday.domain.tag.entity.Tag;
-import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.stereotype.Repository;
-
-import java.util.Optional;
-
-@Repository
-public interface TagRepository extends JpaRepository<Tag, Long> {
-    boolean existsByName(String name);
-}
